revert neard main.rs
